Update README.md
In this example, the check access function takes a user and a resource as arguments and returns True if the user has permission to access the resource, and False otherwise. The permissions matrix defines the access control rules, where each tuple represents a rule that maps a user to a resource and indicates whether the user has permission to access the resource (True) or not (False).